Козлов Михаил. Задание 2. Вариант 11. Вычисление многомерных интегралов с использованием многошаговой схемы (метод Симпсона).

+++ tasks/omp/kozlov_m_simpson_integral/include/ops_omp.hpp
@@ -1,6 +1,8 @@
 // Copyright 2024 Kozlov Mikhail
 #pragma once
 
+#include <omp.h>
+
 #include <cmath>
 #include <functional>
 #include <memory>
@@ -28,7 +30,7 @@ class KozlovTaskSequential : public ppc::core::Task {
  private:
   FUNC f{};
   double x1{}, x2{}, y1{}, y2{};
-  uint64_t n;
-  uint64_t m;
+  int n;
+  int m;
   double res;
 };
